+++ MultiBoard/MultiBoard/Properties/Resources.resx
@@ -149,7 +149,7 @@
     <value>0,1,2,3,4,5,6,7,8,9,A,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P,Q,R,S,T,U,V,W,X,Y,Z,Numpad0,Numpad1,Numpad2,Numpad3,Numpad4,Numpad5,Numpad6,Numpad7,Numpad8,Numpad9,F1,F2,F3,F4,F5,F6,F7,F8,F9,F10,F11,F12,F13,F14,F15,F16,F17,F18,F19,F20,F21,F22,F23,F24</value>
   </data>
   <data name="Version" xml:space="preserve">
-    <value>v1.0.1.6</value>
+    <value>v1.0.1.7-beta</value>
   </data>
   <data name="ChechForUpdatesApiUrl" xml:space="preserve">
     <value>https://api.github.com/repos/Tygo-bear/MultiBoard/releases/latest</value>
